refactor(report): isolate consolidation functions for use outside dot reporters
