+++ src/report/dot/prepareFolderLevel.js
@@ -1,28 +1,9 @@
-const path = require("path").posix;
 const _get = require("lodash/get");
+const consolidateToFolder = require("../utl/consolidateToFolder");
 const moduleUtl = require("./module-utl");
-const consolidateModules = require("./consolidateModules");
-const consolidateModuleDependencies = require("./consolidateModuleDependencies");
-
-function squashDependencyToDir(pDependency) {
-  return {
-    ...pDependency,
-    resolved: path.dirname(pDependency.resolved)
-  };
-}
-
-function squashToDir(pModules) {
-  return pModules.map(pModule => ({
-    ...pModule,
-    source: path.dirname(pModule.source),
-    consolidated: true,
-    dependencies: pModule.dependencies.map(squashDependencyToDir)
-  }));
-}
 
 module.exports = (pResults, pTheme) => {
-  return consolidateModules(squashToDir(pResults.modules))
-    .map(consolidateModuleDependencies)
+  return consolidateToFolder(pResults.modules)
     .sort(moduleUtl.compareOnSource)
     .map(moduleUtl.extractFirstTransgression)
     .map(moduleUtl.folderify)
