+++ src/report/utl/consolidateModuleDependencies.js
